imagebuilder: accept `manifest_type`
The `imagebuilder` phase was missing the `manifest_type` property in its
schema. While pungi (often) guesses correctly for the `manifest_type` it
doesn't do so in the case of ostree installer images; thus it needs to
be allowed.

This was an oversight as the phase implementation already looked for
this value in the configuration.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>